made validateToken func
validateToken handles everything and returns the sub from the jwt. It handles everything so to avoid rewriting a lot of the same code. It returns a somewhat valid error if I want to use the errors more. Every protected endpoint should now start like this:

sub, err := validateToken(ctx)
if err != nil {
         return
}